+++ codex/choices/choices_to_json.py
@@ -8,11 +8,9 @@ from types import MappingProxyType
 
 from caseconverter import camelcase
 
-from codex.choices.admin import (
-    ADMIN_FLAG_CHOICES,
-    ADMIN_TASK_GROUPS,
-)
+from codex.choices.admin import ADMIN_FLAG_CHOICES
 from codex.choices.browser import BROWSER_CHOICES, BROWSER_DEFAULTS
+from codex.choices.jobs import ADMIN_JOBS
 from codex.choices.notifications import Notifications
 from codex.choices.reader import READER_CHOICES, READER_DEFAULTS
 from codex.choices.search import SEARCH_FIELDS
@@ -34,7 +32,7 @@ _DUMPS = MappingProxyType(
 
 _MAP_DUMPS = MappingProxyType(
     {
-        "admin-tasks.json": ADMIN_TASK_GROUPS,
+        "admin-jobs.json": ADMIN_JOBS,
         "browser-defaults.json": BROWSER_DEFAULTS,
         "browser-map.json": BROWSER_CHOICES,
         "reader-defaults.json": READER_DEFAULTS,