+++ frontend/src/plugins/router.js
@@ -10,7 +10,7 @@ const AdminUsersTab = () => import("@/components/admin/tabs/user-tab.vue");
 const AdminGroupsTab = () => import("@/components/admin/tabs/group-tab.vue");
 const AdminLibrariesTab = () =>
   import("@/components/admin/tabs/library-tab.vue");
-const AdminTasksTab = () => import("@/components/admin/tabs/task-tab.vue");
+const AdminJobsTab = () => import("@/components/admin/tabs/job-tab.vue");
 const AdminStatsTab = () => import("@/components/admin/tabs/stats-tab.vue");
 
 const LAST_ROUTE = {
@@ -52,7 +52,7 @@ const routes = [
         component: AdminLibrariesTab,
       },
       { name: "admin-flags", path: "flags", component: AdminFlagsTab },
-      { name: "admin-tasks", path: "tasks", component: AdminTasksTab },
+      { name: "admin-jobs", path: "jobs", component: AdminJobsTab },
       { name: "admin-stats", path: "stats", component: AdminStatsTab },
     ],
   },